requirements.txt
streamlit
transformers
torch
streamlit==1.35.0
requests==2.31.0
python-docx==1.1.0
PyMuPDF==1.24.1
nltk